comments solved, now the mappers are tied
