+++ app/src/main/java/com/example/memeapp/data/service/implementation/MemesServiceImpl.kt
@@ -1,10 +1,9 @@
 package com.example.memeapp.data.service.implementation
 
-import com.example.memeapp.data.mapper.MemesMapper
+import com.example.memeapp.data.mapper.transformListOfMemes
+import com.example.memeapp.data.mapper.transformMemes
 import com.example.memeapp.data.request.generator.MemeRequestGenerator
 import com.example.memeapp.data.service.api.MemeApi
-import com.example.memeapp.data.mapper.MemesMapper.transformListOfMemes
-import com.example.memeapp.data.mapper.MemesMapper.transformMemes
 import com.example.memeapp.domain.entity.MemesEntity
 import com.example.memeapp.domain.service.MemesService
 import io.reactivex.rxjava3.core.Observable
